OP-320: PR fixes
